Final commit before submission
The program works as intended, however...
The program does not use the ChatMessage class that was provided

Might possibly result in a bad grade. Noticed this to close to the
deadline.

There has been updates to every single class.
Unnecessary stuff has been taken out. Synchronization has been thought
of and, measures has been taken to better the incapsulation of methods
and variables.